+++ easing/sine.go
@@ -13,6 +13,10 @@ import (
 
 // ////////////////////////////////////////////////////////////////////////////////// //
 
+const HALF_PI = math.Pi / 2
+
+// ////////////////////////////////////////////////////////////////////////////////// //
+
 // SineIn accelerating from zero velocity
 // https://easings.net/#easeInSine
 func SineIn(t, b, c, d float64) float64 {
@@ -20,7 +24,7 @@ func SineIn(t, b, c, d float64) float64 {
 		return c
 	}
 
-	return -c*math.Cos(t/d*math.Phi) + c + b
+	return -c*math.Cos(t/d*HALF_PI) + c + b
 }
 
 // SineOut decelerating to zero velocity
@@ -30,7 +34,7 @@ func SineOut(t, b, c, d float64) float64 {
 		return c
 	}
 
-	return c*math.Sin(t/d*math.Phi) + b
+	return c*math.Sin(t/d*HALF_PI) + b
 }
 
 // SineInOut acceleration until halfway, then deceleration